% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitBlameRepo.R
\name{blame}
\alias{blame}
\title{blame}
\usage{
blame(repo, file, verbose)
}
\arguments{
\item{repo}{<character> Path to repo}

\item{file}{<character> Relative path to file in repo}

\item{verbose}{<logical> TRUE: Prints messages}
}
\value{
<data.frame>
}
\description{
Workhorse blame function
}
