% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAccuracy.R
\name{calculateAccuracy}
\alias{calculateAccuracy}
\title{A calculate Accuracy Function}
\usage{
calculateAccuracy(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the Accuracy
}
\examples{
calculateAccuracy(2,3,5)
}
\keyword{accuracy}
\keyword{labels,}
\keyword{predictors,}
