% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFscore.R
\name{calculateFscore}
\alias{calculateFscore}
\title{A calculate Fscore Function}
\usage{
calculateFscore(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the Fscore of your data
}
\examples{
calculateFscore(2,3,5)
}
\keyword{Fscore}
\keyword{labels,}
\keyword{predictors,}
