% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNPV.R
\name{calculateNPV}
\alias{calculateNPV}
\title{A calculate NPV Function}
\usage{
calculateNPV(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the NPV of your data
}
\examples{
calculateNPV(2,3,5)
}
\keyword{NPV}
\keyword{labels,}
\keyword{predictors,}
