% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePPV.R
\name{calculatePPV}
\alias{calculatePPV}
\title{A calculate PPV Function}
\usage{
calculatePPV(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the PPV of your data
}
\examples{
calculatePPV(2,3,5)
}
\keyword{PPV}
\keyword{labels,}
\keyword{predictors,}
