% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScoresAtThresholds.R
\name{calculateScoresAtThresholds}
\alias{calculateScoresAtThresholds}
\title{A calculateScoresAtThresholds Function}
\usage{
calculateScoresAtThresholds(labels, predictors,
  metricCalculationFunction = calculateAccuracy, numPermutations = 10)
}
\arguments{
\item{thresholdData}{datapoints main ??? width width of your graph height height of your graph}
}
\description{
This function allows you to calculate scores at various thresholds
}
\examples{
calculateScoresAtThresholds(5,10,calculateAccuracy, 10)
}
\keyword{graphs}
\keyword{thresholdData,}
