% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSensitivity.R
\name{calculateSensitivity}
\alias{calculateSensitivity}
\title{A calculate Sensitivity Function}
\usage{
calculateSensitivity(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the sensitivity of your data
}
\examples{
calculateSensitivity(2,3,5)
}
\keyword{labels,}
\keyword{predictors,}
\keyword{sensitivity}
