% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpecificity.R
\name{calculateSpecificity}
\alias{calculateSpecificity}
\title{A calculate Specificity Function}
\usage{
calculateSpecificity(labels, predictors, threshold)
}
\arguments{
\item{labels}{??? a factor object predictors threshold}
}
\description{
This function allows you to calculate the specificity of your data
}
\examples{
calculateSpecificity(2,3,5)
}
\keyword{labels,}
\keyword{predictors,}
\keyword{specificity}
