% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{checkDependencies}
\alias{checkDependencies}
\title{checkDependencies}
\usage{
checkDependencies(
  repo,
  dependencyType = c("Imports", "Depends"),
  verbose = TRUE
)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}

\item{dependencyType}{<\link[base]{character}> Types of dependencies to be included}

\item{verbose}{<\link[base]{logical}> TRUE or FALSE. If TRUE, progress will be reported.}
}
\value{
<\link[base]{data.frame}> with all the packages that are now permitted.
}
\description{
Check package dependencies
}
