% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies_OLD.R
\name{checkDependenciesOLD}
\alias{checkDependenciesOLD}
\title{checkDependencies}
\usage{
checkDependenciesOLD(
  packageName = NULL,
  dependencyType = c("Imports", "Depends")
)
}
\arguments{
\item{packageName}{Name of package to profile. If NULL current package}

\item{dependencyType}{Imports, depends, and/ or suggests}
}
\value{
Returns value NULL
}
\description{
[DEPRECATED] Check package dependencies
}
\examples{
# Run only in interactive session
if (interactive()) {
  checkDependencies(packageName = "PaRe")

  checkDependencies(packageName = "PaRe", c("Suggests"))
}
}
