% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-countPackageLines.R, R/countPackageLines.R
\name{countLines}
\alias{countLines}
\title{countLines}
\usage{
countLines(files)

countLines(files)
}
\arguments{
\item{files}{List of files}
}
\value{
numeric

numeric
}
\description{
Counts the lines of a list of files

Counts the lines of a list of files
}
