% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPackageLines.R
\name{countPackageLines}
\alias{countPackageLines}
\title{countPackageLines}
\usage{
countPackageLines(repo)
}
\arguments{
\item{repo}{(`Repository`)
Repository object.}
}
\value{
(`data.frame()`)
  Tibble containing the amount of lines per file in the Repository object.
}
\description{
countPackageLines
}
