% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{forward}
\alias{forward}
\title{forward}
\usage{
forward(c, nodes)
}
\arguments{
\item{c}{<\link[base]{character}> Formatted function name.}

\item{nodes}{<\link[base]{c}> Vector of all individual paths.}
}
\value{
<\link[base]{vector}> Vector of nodes forward in the graph.
}
\description{
forward
}
