% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{functionUseGraph}
\alias{functionUseGraph}
\title{functionUseGraph}
\usage{
functionUseGraph(repo)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> Repository object.}
}
\value{
<\link[igraph]{graph}> Graph object crated by \link[igraph]{graph_from_data_frame}.
}
\description{
functionUseGraph
}
