% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{funDiagram}
\alias{funDiagram}
\title{funDiagram}
\usage{
funDiagram(funName, nodes, maxIter, n = 0)
}
\arguments{
\item{funName}{<\link[base]{character}> Name of function to get path of.}

\item{nodes}{<\link[base]{c}> Vector of all individual paths.}

\item{maxIter}{<\link[base]{numeric}> Maximum iterations specified by user.}

\item{n}{<\link[base]{numeric}> (0) Internal counter of n iterations}
}
\value{
Diagram `htmlwidget` object. See \link[htmlwidgets]{createWidget}
}
\description{
funDiagram
}
