% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{funsUsedInLine}
\alias{funsUsedInLine}
\title{funsUsedInLine}
\usage{
funsUsedInLine(lines, name, i, verbose = FALSE)
}
\arguments{
\item{lines}{<\link[base]{c}> of <\link[base]{character}>}

\item{name}{<\link[base]{character}>}

\item{i}{<\link[base]{numeric}> line}

\item{verbose}{<\link[base]{logical}> Prints message when no function found}
}
\value{
<\link[base]{data.frame}> of 3 colums: Package (pkg); Function (fun); Line
in script (line)
}
\description{
Support function for funsUsedInFile.
}
