% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getApplyCall}
\alias{getApplyCall}
\title{getApplyCall}
\usage{
getApplyCall(fun, defFuns)
}
\arguments{
\item{fun}{<\link[PaRe]{Function}> Function object.}

\item{defFuns}{<\link[base]{data.frame}>}
}
\value{
<\link[base]{data.frame}>
}
\description{
getApplyCall
}
