% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getApplyFromLines}
\alias{getApplyFromLines}
\title{getApplyFromLines}
\usage{
getApplyFromLines(lines)
}
\arguments{
\item{lines}{(\link[base]{c}) Vector of (\link[base]{character}).
See \link[PaRe]{getDefinedFunctions}}
}
\value{
(\link[base]{character}).
}
\description{
getApplyFromLines
}
