% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getApplyFun}
\alias{getApplyFun}
\title{getApplyFun}
\usage{
getApplyFun(line, lines)
}
\arguments{
\item{line}{<\link[base]{numeric}> Current line number.}

\item{lines}{<\link[base]{c}> of <\link[base]{character}> lines.}
}
\value{
<\link[base]{c}> of <\link[base]{character}> of functions found in
\link[base]{apply} function call.
}
\description{
getApplyFun
}
