% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefinedFunctions.R
\name{getBodyIndices}
\alias{getBodyIndices}
\title{getBodyIndices}
\usage{
getBodyIndices(line, lines)
}
\arguments{
\item{line}{Line index of the function constructor.}

\item{lines}{All lines of the R-file to be investigated.}
}
\value{
Returns a data.frame with the start and end indices of lines.
}
\description{
Helper function for getDefinedFunctions, retrieves offset indeces for where
the body of the function starts.
}
