% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranPackage.R
\name{getCranPackage}
\alias{getCranPackage}
\title{getCranPackage}
\usage{
getCranPackage(pkgPath)
}
\arguments{
\item{pkgPath}{package path}
}
\value{
CranPackage or NULL if it does not exist
}
\description{
get CranPackage object from pkgPath. Returns NULL if there is no CRAN package.
}
