% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefinedFunctions.R
\name{getDefinedFunctionsFile}
\alias{getDefinedFunctionsFile}
\title{getDefinedFunctionsFile}
\usage{
getDefinedFunctionsFile(filePath, verbose = FALSE)
}
\arguments{
\item{filePath}{File path to the R-file to be investigated}

\item{verbose}{Prints message as to what file is currently being worked on.
Usefull if used in an apply funciton, investigating alot of different files.}
}
\value{
Returns a tibble object.
}
\description{
Gets all the defined functions in a file, stored in a a tibble, with the
following columns: file (filename), start (start line in file), size (Amount
of lines occupied by function), fun (function name).
}
\examples{
filePath <- system.file(package = "PaRe", "testScript.R")
df <- getDefinedFunctionsFile(filePath)
}
