% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefinedFunctions.R
\name{getDefinedFunctionsPkg}
\alias{getDefinedFunctionsPkg}
\title{getDefinedFunctionsPkg}
\usage{
getDefinedFunctionsPkg(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to package}

\item{verbose}{Verbose messages}
}
\value{
tibble
}
\description{
Gets defined functions of the package
}
