% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies_OLD.R
\name{getDiffVersions}
\alias{getDiffVersions}
\title{getDiffVersions}
\usage{
getDiffVersions(dependencies, permittedPackages)
}
\arguments{
\item{dependencies}{Dependencies}

\item{permittedPackages}{permittedPackages}
}
\value{
versions of permitted packages
}
\description{
[DEPRECATED] Helper function
}
