% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getDlply}
\alias{getDlply}
\title{getDlply}
\usage{
getDlply(line, lines)
}
\arguments{
\item{line}{<\link[base]{numeric}> Current line number.}

\item{lines}{<\link[base]{c}> of <\link[base]{character}> lines.}
}
\value{
<\link[base]{c}> of <\link[base]{character}> of functions found in
\link[plyr]{dlply} function call.
}
\description{
getDlply
}
