% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getDlplyCallFromLines}
\alias{getDlplyCallFromLines}
\title{getDlplyCallFromLines}
\usage{
getDlplyCallFromLines(lines)
}
\arguments{
\item{lines}{(\link[base]{c}) Vector of (\link[base]{character}).}
}
\value{
(\link[base]{character}).
}
\description{
getDlplyCallFromLines
}
