% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getDoCallFromLines}
\alias{getDoCallFromLines}
\title{getDoCallFromLines}
\usage{
getDoCallFromLines(lines)
}
\arguments{
\item{lines}{<\link[base]{c}> Vector of <\link[base]{character}>.}
}
\value{
<\link[base]{character}>.
}
\description{
getDoCallFromLines
}
