% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getFunCall}
\alias{getFunCall}
\title{getFunCall}
\usage{
getFunCall(fun, defFuns)
}
\arguments{
\item{fun}{(\link[PaRe]{Function}) Function object.}

\item{defFuns}{(\link[base]{data.frame})
See \link[PaRe]{getDefinedFunctions}.}
}
\value{
(\link[base]{data.frame})
}
\description{
getFunCall
}
