% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getFunctionUse}
\alias{getFunctionUse}
\title{summariseFunctionUse}
\usage{
getFunctionUse(repo, verbose = FALSE)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}

\item{verbose}{<\link[base]{logical}> Default: FALSE; prints message to console which file is
currently being worked on.}
}
\value{
<\[dplyr]{tibble}>
}
\description{
Summarise functions used in R package
}
