% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getFunsPerDefFun}
\alias{getFunsPerDefFun}
\title{Title}
\usage{
getFunsPerDefFun(files, defFuns)
}
\arguments{
\item{files}{<\link[base]{list}> of <\link[PaRe]{File}> objects.}

\item{defFuns}{<\link[base]{data.frame}>}
}
\value{
<\link[base]{data.frame}>
}
\description{
Title
}
