% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphData.R
\name{getGraphData}
\alias{getGraphData}
\title{getGraphData}
\usage{
getGraphData(repo, packageTypes = c("Imports", "Suggests"))
}
\arguments{
\item{repo}{\link[PaRe]{Repository} object.}

\item{packageTypes}{<\link[base]{c}> of <\link[base]{character}> Types of packages to be
included in the result. Default: c("imports", "depends") Available types
are: "imports", "depends", "suggests", "enhances", "linkingto"}
}
\value{
<\link[tidygraph]{as_tbl_graph}>
}
\description{
getGraphData
}
