% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halstead.R
\name{getHalstead}
\alias{getHalstead}
\title{getHalstead}
\usage{
getHalstead(fun)
}
\arguments{
\item{fun}{\link[PaRe]{Function}\cr
Function object.}
}
\value{
\link[base]{data.frame}
Data frame with the following columns:\tabular{ll}{
   column \tab data type \cr
   funName \tab \link[base]{character} \cr
   vocabulary \tab \link[base]{integer} \cr
   progLength \tab \link[base]{integer} \cr
   callEstProgLength \tab \link[base]{numeric} \cr
   volume \tab \link[base]{numeric} \cr
   difficulty \tab \link[base]{numeric} \cr
   effort \tab \link[base]{numeric} \cr
}
}
\description{
Halstead complexity measures are measureable properties of a piece of code.
\cr\cr
Number of Operators: \eqn{N_1}\cr
Number of Operands: \eqn{N_2}\cr
Total number of Operators: \eqn{\eta_1}\cr
Total number of Operands: \eqn{\eta_2}
\cr\cr
Program Vocabulary (vocabulary):\cr
\eqn{V = \eta_1 + \eta_2}
\cr\cr
Program Length (progLength):\cr
\eqn{N = N_1 + N_2}
\cr\cr
Calculated estimated program length  (calEstProgLength):\cr
\eqn{\hat{N} = \eta_1 \times log_2\ \eta_1 + \eta_2 \times log_2\ \eta_2}
\cr\cr
Volume (volume):\cr
\eqn{V = N \times log_2\ \eta}
\cr\cr
Difficulty (difficulty):\cr
\eqn{D = \frac{\eta_1}{2} \times \frac{N_2}{\eta_2}}
\cr\cr
Effort (effort):\cr
\eqn{E = D \times V}
\cr\cr
\href{\verb{https://www.researchgate.net/publication/235978157_Elements_Of_Software_Science}}{Elements Of Software Science, Halstead M.H (1977)}
\cr
\href{\verb{https://en.wikipedia.org/wiki/Halstead_complexity_measures}}{Wiki}
}
