% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInvestigators.R
\name{getMultiLineFun}
\alias{getMultiLineFun}
\title{getMultiLineFun}
\usage{
getMultiLineFun(line, lines)
}
\arguments{
\item{line}{(\link[base]{numeric})\cr
Current line number.}

\item{lines}{(\link[base]{c})\cr
Vector of (\link[base]{character}) lines.}
}
\value{
(\link[base]{character})
}
\description{
getMultiLineFun
}
