% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{getNotPermitted}
\alias{getNotPermitted}
\title{getNotPermitted}
\usage{
getNotPermitted(dependencies, permittedPackages)
}
\arguments{
\item{dependencies}{Dependencies}

\item{permittedPackages}{Packages that are permitted as character vector}
}
\value{
Returns vector of not permitted packages
}
\description{
Helper function
}
