% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitBlameRepo.R
\name{gitBlameRepo}
\alias{gitBlameRepo}
\title{gitBlameRepo}
\usage{
gitBlameRepo(repoPath, regex = "\\\\.[Rr]$", verbose = FALSE, threads = 0)
}
\arguments{
\item{repoPath}{<character> Path to repository.}

\item{regex}{<character> Reg-ex pattern to use to detect files.}

\item{verbose}{<logical> TRUE: Prints status messages.}

\item{threads}{<numeric> Number of threads to use. Useful for a large quantity of files.}
}
\value{
<data.frame>
}
\description{
Retrieves a data.frame of authors per file.
}
