% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefinedFunctions.R
\name{goToBody}
\alias{goToBody}
\title{goToBody}
\usage{
goToBody(line, lines)
}
\arguments{
\item{line}{Line number of the constructor of the function.}

\item{lines}{Lines of the R-file to be investigated.}
}
\value{
Returns a numeric index.
}
\description{
Helper function for getBodyIndices and getDefinedFunctions. Computes the
starting index of the function body.
}
