% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintFile}
\alias{lintFile}
\title{lintFile}
\usage{
lintFile(fileName)
}
\arguments{
\item{fileName}{Path to file to lint}
}
\value{
list of lint objects.
}
\description{
Lint a given file.
}
\examples{
lintFile(
  fileName = system.file(package = "PaRe", "testScript.R")
)
}
