% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintPackage}
\alias{lintPackage}
\title{lintPackage}
\usage{
lintPackage(path = ".")
}
\arguments{
\item{path}{Default: ".", Path to package}
}
\value{
List of lint objects.
}
\description{
Lintr object, using default lintr object with camelCase
}
\examples{
lintPackage()
}
