% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintScore}
\alias{lintScore}
\title{lintScore}
\usage{
lintScore(lintFunction, ...)
}
\arguments{
\item{lintFunction}{Lint function to use}

\item{...}{Other parameters a Lint function might need (i.e. file name)}
}
\value{
A tibble of percentage scores per type of Lint message.
}
\description{
Function that scores the lintr output as a percentage per message type
(style, warning, error). Lintr messages / lines assessed * 100
}
\examples{
# With file lintr
lintScore(
  lintFunction = lintFile,
  system.file(package = "PaRe", "testScript.R")
)

# With standard package lintr
lintScore(
  lintFunction = lintr::lint_package,
  system.file(package = "PaRe")
)
}
