% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies_OLD.R
\name{messagePackageVersion}
\alias{messagePackageVersion}
\title{messagePackageVersion}
\usage{
messagePackageVersion(i, diffVersions)
}
\arguments{
\item{i}{iterator}

\item{diffVersions}{different versions}
}
\description{
[DEPRECATED] Helper message function
}
