% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies_OLD.R
\name{messagePermission}
\alias{messagePermission}
\title{messagePermission}
\usage{
messagePermission(i, not_permitted)
}
\arguments{
\item{i}{iterator}

\item{not_permitted}{Not permitted}
}
\description{
Helper message function
}
