% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaRe-package.R
\docType{package}
\name{PaRe-package}
\alias{PaRe}
\alias{PaRe-package}
\title{PaRe: PaRe (Package Reviewer) is the Successor of the DependencyReviewer Package}
\description{
Reviews other packages during code review by looking at their dependencies, code style, code complexity, and how interanlly defined functions interact with one another.
}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl}

}
\keyword{internal}
