% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pare.R
\name{pare}
\alias{pare}
\title{PAREs: PARtial Embeddings for deconfounded dimension reduction}
\usage{
pare(Y, formula, data = NULL, embed, type = c("MDS", "dist"), ...)
}
\arguments{
\item{Y}{Input data or distance matrix. If \code{Y} does not have class "dist",
then Euclidean distance is used to construct the distance matrix}

\item{formula}{Formula for regressing out confounding variables}

\item{data}{Data frame containing variables in the model specified by \code{formula}}

\item{embed}{Dimension reduction method. Must be a function such as
\link[Rtsne]{Rtsne} or \link[umap]{umap}}

\item{type}{Type of input into \code{embed}. Refer to documentation of embedding
function}

\item{...}{Additional arguments passed to \code{embed}}
}
\value{
\code{pare} returns the output of \code{embed} and the format depends on the
method specified
}
\description{
\code{pare} implements a framework for removing nuisance covariates from any
distance-based dimension reduction method including \link[Rtsne]{Rtsne} and
\link[umap]{umap}. The \code{pare} function takes either data or a distance matrix
as an input then passes a deconfounded distance matrix to a dimension
reduction method. The output is a PARE with respect to the specified
confounder.
}
\examples{
pare(iris[,-5], ~ Species, data = iris, cmdscale, type = "dist")
}
