% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareWhiteSpace.R
\name{pareWhiteSpace}
\alias{pareWhiteSpace}
\title{pareWhiteSpace}
\usage{
pareWhiteSpace(filePath, outPath = NULL)
}
\arguments{
\item{filePath}{Path to file.}

\item{outPath}{Default NULL: Path to output file. If NULL file will be
overwritten.}
}
\description{
pareWhiteSpace
}
