% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareWhiteSpace.R
\name{pareWhiteSpacePkg}
\alias{pareWhiteSpacePkg}
\title{pareWhiteSpacePkg}
\usage{
pareWhiteSpacePkg(pkgPath = "./", overwrite = TRUE)
}
\arguments{
\item{pkgPath}{Path to package}

\item{overwrite}{Default: TRUE boolean to overwrite original files or not.}
}
\description{
pareWhiteSpacePkg
}
