% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{pkgDiagram}
\alias{pkgDiagram}
\title{pkgDiagram}
\usage{
pkgDiagram(repo, verbose = FALSE, ...)
}
\arguments{
\item{verbose}{(`logical()`)
Verbose messages}

\item{...}{Optional other parameters for \link[DiagrammeR]{grViz}.}

\item{pkgPath}{(`character()`)
Path to package}
}
\value{
`htmlwidget`
  Diagram `htmlwidget` object. See \link[htmlwidgets]{createWidget}
}
\description{
Creates a diagram of all defined functions in a package.
}
