% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPareScores.R
\name{plotPareScores}
\alias{plotPareScores}
\title{A plotPareScores Function}
\usage{
plotPareScores(accuracyThresholdData, metrics, ylab = "Accuracy", main = "")
}
\arguments{
\item{thresholdData}{datapoints main ??? width width of your graph height height of your graph}
}
\description{
This function allows you to plot PARE scores using the ggplot2 package
}
\examples{
plotPareScores(thresholdData, main = "", width=5, height=100)
}
\keyword{datapoints}
\keyword{graphs,}
\keyword{thresholdData,}
