% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPareThresholdSelection.R
\name{plotPareThresholdSelection}
\alias{plotPareThresholdSelection}
\title{A plotPareThresholdSelection Function}
\usage{
plotPareThresholdSelection(accuracyThresholdData, main = "")
}
\arguments{
\item{thresholdData}{datapoints main ??? width width of your graph height height of your graph}
}
\description{
This function allows you to plot PARE Threshold Selection ?? using the ggplot2 package
}
\examples{
plotPareScores(thresholdData, main = "", width=5, height=100)
}
\keyword{graphs}
\keyword{thresholdData,}
