% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-checkDependencies.R
\name{R6checkDependencies}
\alias{R6checkDependencies}
\title{checkDependencies}
\usage{
R6checkDependencies(
  repo,
  dependencyType = c("Imports", "Depends"),
  verbose = TRUE
)
}
\arguments{
\item{dependencyType}{Types of dependencies to be included}

\item{verbose}{TRUE or FALSE. If TRUE, progress will be reported.}

\item{pkgPath}{Path to package}
}
\value{
Returns a data.frame with all the packages that are now permitted.
}
\description{
Check package dependencies
}
