% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{recBackward}
\alias{recBackward}
\title{recBackward}
\usage{
recBackward(c, nodes, n, maxIter)
}
\arguments{
\item{c}{<\link[base]{character}> Formatted function name.}

\item{nodes}{<\link[base]{c}> Vector of all individual paths.}

\item{n}{<\link[base]{numeric}> Internal counter of n iterations}

\item{maxIter}{<\link[base]{numeric}> Maximum iterations specified by user.}
}
\value{
<\link[base]{character}>
}
\description{
recBackward
}
