% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseFunctionUse.R
\name{summariseFunctionUse}
\alias{summariseFunctionUse}
\title{summariseFunctionUse}
\usage{
summariseFunctionUse(r_files, verbose = FALSE)
}
\arguments{
\item{r_files}{Complete path(s) to files to be investigated}

\item{verbose}{Default: FALSE; prints message to console which file is
currently being worked on.}
}
\value{
tibble
}
\description{
Summarise functions used in R package
}
\examples{
summariseFunctionUse(
  r_files = system.file(package = "PaRe", "testScript.R"))

# Only in an interactive session
if (interactive()) {
  summariseFunctionUse(list.files(here::here("R"), full.names = TRUE))
}
}
