% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaRe-package.R
\docType{data}
\name{whiteList}
\alias{whiteList}
\title{whiteList}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\usage{
whiteList
}
\description{
data.frame containing links to csv-files which should be used to fetch
white-listed dependencies.
}
\details{
By default three csv's are listed:
1. darwin
2. hades
3. tidyverse

The data.frame is locally fetched under:
`system.file(package = "PaRe", "whiteList.csv")`

Manual insertions into this data.frame can be made, or the data.frame can
be overwritten entirely.

The data.frame itself has the following structure:
`data.frame(source = c(), link = c(), package = c(), version = c())`
`source` is the name of the source
`link` is the link or path to the csv-file
`package` is the columnname of the package name column in the csv-file being
linked to.
`version` is the columnname of the version column in the csv-file being
linked to.

The csv-files that are being pointed to should have the following structure:
}
\examples{
if (interactive()) {
  # Dropping tidyverse
  whiteList <- whiteList \%>\%
    dplyr::filter(source != "tidyverse")

  # getDefaultPermittedPackages will now only use darwin and hades
  getDefaultPermittedPackages()
}
}
\keyword{datasets}
