% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{checkDependencies}
\alias{checkDependencies}
\title{checkDependencies}
\usage{
checkDependencies(
  pkgPath = "./",
  dependencyType = c("Imports", "Depends"),
  verbose = TRUE
)
}
\arguments{
\item{pkgPath}{Path to package}

\item{dependencyType}{Types of dependencies to be included}

\item{verbose}{TRUE or FALSE. If TRUE, progress will be reported.}
}
\value{
Returns a data.frame with all the packages that are now permitted.
}
\description{
Check package dependencies
}
\examples{
if (interactive()) {
  checkDependencies(system.file(package = "PaRe", "glue"))
}
}
