% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{checkInstalled}
\alias{checkInstalled}
\title{checkInstalled}
\usage{
checkInstalled()
}
\value{
<\link[base]{logical}> Boolean depending if suggested packages are installed.
}
\description{
Checks if suggested packages are installed.
}
