% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPackageLines.R
\name{countPackageLines}
\alias{countPackageLines}
\title{countPackageLines}
\usage{
countPackageLines(
  repo,
  fileEx = c("R", "cpp", "sql", "java"),
  ignoreDirs = c("tests", "extras")
)
}
\arguments{
\item{fileEx}{File extensions to search for, is case sensitive.}

\item{ignoreDirs}{Directories to ignore when searching for files.}

\item{path}{Path to package}
}
\value{
Tibble
}
\description{
Counts the lines of files ending with a specific extension.
}
\examples{
countPackageLines("./")
}
