% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-File.R
\name{File}
\alias{File}
\title{R6 File class}
\description{
Class representing a file containing code.
}
\section{Super class}{
\code{\link[PaRe:Code]{PaRe::Code}} -> \code{File}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-File-new}{\code{File$new()}}
\item \href{#method-File-getFunctions}{\code{File$getFunctions()}}
\item \href{#method-File-getFunctionTable}{\code{File$getFunctionTable()}}
\item \href{#method-File-getType}{\code{File$getType()}}
\item \href{#method-File-clone}{\code{File$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PaRe" data-topic="Code" data-id="getLines"><a href='../../PaRe/html/Code.html#method-Code-getLines'><code>PaRe::Code$getLines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PaRe" data-topic="Code" data-id="getNLines"><a href='../../PaRe/html/Code.html#method-Code-getNLines'><code>PaRe::Code$getNLines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PaRe" data-topic="Code" data-id="getName"><a href='../../PaRe/html/Code.html#method-Code-getName'><code>PaRe::Code$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PaRe" data-topic="Code" data-id="print"><a href='../../PaRe/html/Code.html#method-Code-print'><code>PaRe::Code$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-new"></a>}}
\if{latex}{\out{\hypertarget{method-File-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{<\link[base]{character}> Path to file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`invisible(self)`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-getFunctions"></a>}}
\if{latex}{\out{\hypertarget{method-File-getFunctions}{}}}
\subsection{Method \code{getFunctions()}}{
Get method to get a list of Function objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$getFunctions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<\link[base]{list}> of \link[PaRe]{Function} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-getFunctionTable"></a>}}
\if{latex}{\out{\hypertarget{method-File-getFunctionTable}{}}}
\subsection{Method \code{getFunctionTable()}}{
Get method to retrieve the function table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$getFunctionTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<\link[base]{data.frame}>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-getType"></a>}}
\if{latex}{\out{\hypertarget{method-File-getType}{}}}
\subsection{Method \code{getType()}}{
Gets type of file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$getType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
<\link[base]{character}>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-clone"></a>}}
\if{latex}{\out{\hypertarget{method-File-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
