% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-getFunctionUse.R, R/summariseFunctionUse.R
\name{funsUsedInFile}
\alias{funsUsedInFile}
\title{funsUsedInFile}
\usage{
funsUsedInFile(files, verbose = FALSE)

funsUsedInFile(files, verbose = FALSE)
}
\arguments{
\item{files}{Files to get functions from}

\item{verbose}{Verbosity}
}
\value{
table

table
}
\description{
Support function

Support function
}
