% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{funsUsedInFile}
\alias{funsUsedInFile}
\title{funsUsedInFile}
\usage{
funsUsedInFile(files, verbose = FALSE)
}
\arguments{
\item{files}{<\link[base]{list}> of <\link[PaRe]{File}> objects.}

\item{verbose}{<\link[base]{logical}>}
}
\value{
<\link[base]{list}>
}
\description{
Support function
}
