% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{funsUsedInLine}
\alias{funsUsedInLine}
\title{funsUsedInLine}
\usage{
funsUsedInLine(lines, name, i, verbose = FALSE)
}
\arguments{
\item{lines}{(\link[base]{c}) of (\link[base]{character})\cr}

\item{name}{(\link[base]{character})\cr}

\item{i}{(\link[base]{numeric})\cr}

\item{verbose}{(\link[base]{logical}: FALSE)\cr}
}
\value{
(\link[base]{data.frame})\cr
\enumerate{
  \item (\link[base]{character}) pkg
  \item (\link[base]{character}) fun
  \item (\link[base]{numeric}) line
}
}
\description{
Support function for funsUsedInFile.
}
