% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-getFunctionUse.R, R/summariseFunctionUse.R
\name{funsUsedInLine}
\alias{funsUsedInLine}
\title{funsUsedInLine}
\usage{
funsUsedInLine(file_txt, file_name, i, verbose = FALSE)

funsUsedInLine(file_txt, file_name, i, verbose = FALSE)
}
\arguments{
\item{file_txt}{file to use}

\item{file_name}{name of file}

\item{i}{line}

\item{verbose}{Prints message when no function found}
}
\value{
data.frame of 3 colums: Package (pkg); Function (fun); Line in
script (line)

data.frame of 3 colums: Package (pkg); Function (fun); Line in
script (line)
}
\description{
Support function for funsUsedInFile.

Support function for funsUsedInFile.
}
