% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{funsUsedInLine}
\alias{funsUsedInLine}
\title{funsUsedInLine}
\usage{
funsUsedInLine(lines, name, i, verbose = FALSE)
}
\arguments{
\item{lines}{(\link[base]{c}) of (\link[base]{character})}

\item{name}{(\link[base]{character})}

\item{i}{(\link[base]{numeric})}

\item{verbose}{(\link[base]{logical}: FALSE)}
}
\value{
(\link[base]{data.frame})\tabular{ll}{
   column \tab data type \cr
   pkg \tab \link[base]{character} \cr
   fun \tab \link[base]{character} \cr
   line \tab \link[base]{numeric} \cr
}
}
\description{
Support function for funsUsedInFile.
}
