% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{funsUsedInLine}
\alias{funsUsedInLine}
\title{funsUsedInLine}
\usage{
funsUsedInLine(lines, name, i, verbose = FALSE)
}
\arguments{
\item{i}{line}

\item{verbose}{Prints message when no function found}

\item{file_txt}{file to use}

\item{file_name}{name of file}
}
\value{
data.frame of 3 colums: Package (pkg); Function (fun); Line in
script (line)
}
\description{
Support function for funsUsedInFile.
}
