% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getApplyFun}
\alias{getApplyFun}
\title{getApplyFun}
\usage{
getApplyFun(line, lines)
}
\arguments{
\item{line}{All lines}

\item{lines}{Current line number}
}
\value{
Returns function in (vsl)apply function call.
}
\description{
getApplyFun
}
