% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPermittedPackages.R
\name{getDefaultPermittedPackages}
\alias{getDefaultPermittedPackages}
\title{getDefaultPermittedPackages}
\usage{
getDefaultPermittedPackages(base = TRUE)
}
\arguments{
\item{base}{Boolean if base package should be included}
}
\value{
tibble of two columns (package, version) with all 'allowed'
packages.
}
\description{
Gets permitted packages
}
\examples{
# Run only in interactive session
if (interactive()) {
  getDefaultPermittedPackages()
}

}
