% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPermittedPackages.R
\name{getDefaultPermittedPackages}
\alias{getDefaultPermittedPackages}
\title{getDefaultPermittedPackages}
\usage{
getDefaultPermittedPackages(base = TRUE)
}
\arguments{
\item{base}{<\link[base]{logical}> if base package should be included}
}
\value{
<\link[dplyr]{tibble}> of two columns (package, version) with all 'allowed'
packages.
}
\description{
Gets permitted packages
}
