% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPermittedPackages.R
\name{getDefaultPermittedPackages}
\alias{getDefaultPermittedPackages}
\title{getDefaultPermittedPackages}
\usage{
getDefaultPermittedPackages(nThreads = 1)
}
\arguments{
\item{nThreads}{(\code{numeric(1)}: 1)
Number of threads to use to fetch permitted packages}
}
\value{
(\link[dplyr]{tibble})\tabular{ll}{
   column \tab data type \cr
   package \tab \link[base]{character} \cr
   version \tab \link[base]{character} \cr
}
}
\description{
Gets permitted packages. An internet connection is required.
}
\examples{
\donttest{
# Set cache
withr::local_envvar(
  R_USER_CACHE_DIR = tempfile()
)

if (interactive()) {
  getDefaultPermittedPackages()
}
}
}
