% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPermittedPackages.R
\name{getDefaultPermittedPackages}
\alias{getDefaultPermittedPackages}
\title{getDefaultPermittedPackages}
\usage{
getDefaultPermittedPackages(base = TRUE)
}
\arguments{
\item{base}{(\link[base]{logical}: TRUE)\cr
\describe{
  \item{TRUE}{Base packages will be included.}
  \item{FALSE}{Base packages will be ignored.}
}}
}
\value{
(\link[dplyr]{tibble})\cr
\enumerate{
  \item (\link[base]{character}) package
  \item (\link[base]{character}) version
}
}
\description{
Gets permitted packages. An internet connection is required.
}
\examples{
# Set cache
withr::local_envvar(
  R_USER_CACHE_DIR = tempfile()
)

getDefaultPermittedPackages()
}
