% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefinedFunctions.R
\name{getDefinedFunctions}
\alias{getDefinedFunctions}
\title{getDefinedFunctions}
\usage{
getDefinedFunctions(repo)
}
\arguments{
\item{repo}{(`PaRe::Repository`)
\link[PaRe]{Repository} object}
}
\value{
(`data.frame()`)
data.frame
}
\description{
getDefinedFunctions
}
