% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getDlply}
\alias{getDlply}
\title{getDlply}
\usage{
getDlply(line, lines)
}
\arguments{
\item{line}{All lines}

\item{lines}{Current line number}
}
\value{
Returns function in plyr::dlply function call.
}
\description{
getDlply
}
