% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-getFunctionUse.R, R/summariseFunctionUse.R
\name{getDoCallFun}
\alias{getDoCallFun}
\title{getDoCallFun}
\usage{
getDoCallFun(line, lines)

getDoCallFun(line, lines)
}
\arguments{
\item{line}{Current line number}

\item{lines}{All lines}
}
\value{
Returns function used in do.call function call.

Returns function used in do.call function call.
}
\description{
getDoCallFun

getDoCallFun
}
