% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getExportedFunctions}
\alias{getExportedFunctions}
\title{getExportedFunctions}
\usage{
getExportedFunctions(path)
}
\arguments{
\item{path}{(\link[base]{character})\cr
Path to package}
}
\value{
(\link[base]{c})
Vector of \link[base]{character}  exported functions.
}
\description{
Gets all the exported functions of a package, from NAMESPACE.
}
