% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getExportedFunctions}
\alias{getExportedFunctions}
\title{getExportedFunctions}
\usage{
getExportedFunctions(path)
}
\arguments{
\item{path}{(`character()`)
Path to package}
}
\value{
(`c()`) of (`character()`)
  vector of exported functions
}
\description{
Gets all the exported functions of a package, from NAMESPACE.
}
