% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{getFunctionDiagram}
\alias{getFunctionDiagram}
\title{subsetGraph}
\usage{
getFunctionDiagram(repo, functionName)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> Repository object.}

\item{functionName}{<\link[base]{character}> Name of function.}
}
\value{
`htmlwidgets` Substted diagram. See \link[DiagrammeR]{grViz}
}
\description{
subsetGraph
}
