% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{getFunctionDiagram}
\alias{getFunctionDiagram}
\title{getFunctionDiagram}
\usage{
getFunctionDiagram(repo, funName, maxIter = 10)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}

\item{funName}{<\link[base]{character}> Name of function to get path of.}

\item{maxIter}{<\link[base]{numeric}> (10) Maximum iterations specified by user.}
}
\value{
Diagram `htmlwidget` object. See \link[htmlwidgets]{createWidget}
}
\description{
getFunctionDiagram
}
