% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionUse.R
\name{getFunctionUse}
\alias{getFunctionUse}
\title{summariseFunctionUse}
\usage{
getFunctionUse(files, verbose = FALSE)
}
\arguments{
\item{verbose}{Default: FALSE; prints message to console which file is
currently being worked on.}

\item{r_files}{Complete path(s) to files to be investigated}
}
\value{
tibble
}
\description{
Summarise functions used in R package
}
