% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getFunsPerDefFun}
\alias{getFunsPerDefFun}
\title{getFunsPerDefFun}
\usage{
getFunsPerDefFun(files, allFuns, verbose)
}
\arguments{
\item{files}{Vector of files to investigate.}

\item{allFuns}{allFunctions data.frame}

\item{verbose}{Verbose messages}
}
\value{
returns data.frame of all functions per defined function of package.
}
\description{
Gets all function calls per defined function in the package.
}
