% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getFunsPerDefFun}
\alias{getFunsPerDefFun}
\title{getFunsPerDefFun}
\usage{
getFunsPerDefFun(files, allFuns, verbose)
}
\arguments{
\item{files}{(`list()`)
List of files to investigate.}

\item{allFuns}{(`data.frame()`)
Data.frame of allFunctions.}

\item{verbose}{(`logical()`)
Logical to turn verbose messaging on/off.}
}
\value{
(`data.frame()`)
  Data.frame of all functions per defined function of package.
}
\description{
Gets all function calls per defined function in the package.
}
