% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{getFunsPerDefFun}
\alias{getFunsPerDefFun}
\title{getFunsPerDefFun}
\usage{
getFunsPerDefFun(files, allFuns, verbose)
}
\arguments{
\item{files}{(`list()`)
<\link[base]{list}> of <\link[PaRe]{File}> objects to investigate.}

\item{allFuns}{<\link[base]{data.frame}> of allFunctions.}

\item{verbose}{<\link[base]{logical}> to turn verbose messaging on/off.}
}
\value{
<\link[base]{data.frame}> of all functions per defined function of package.
}
\description{
Gets all function calls per defined function in the package.
}
