% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphData.R
\name{getGraphData}
\alias{getGraphData}
\title{getGraphData}
\usage{
getGraphData(
  path = "./",
  excluded_packages = c(""),
  package_types = c("imports", "depends")
)
}
\arguments{
\item{path}{Path to package}

\item{excluded_packages}{Packages to exclude}

\item{package_types}{Types of packages to be included in the result. Default: c("imports", "depends")
Available types are: "imports", "depends", "suggests", "enhances", "linkingto"}
}
\value{
net_data graph data
}
\description{
getGraphData
}
\examples{
# Only run in interactive session
if (interactive()) {
  graphData <- getGraphData()
}
}
