% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphData.R
\name{getGraphData}
\alias{getGraphData}
\title{getGraphData}
\usage{
getGraphData(
  repo,
  excludePackages = c(""),
  packageTypes = c("imports", "depends")
)
}
\arguments{
\item{packageTypes}{Types of packages to be included in the result. Default: c("imports", "depends")
Available types are: "imports", "depends", "suggests", "enhances", "linkingto"}

\item{path}{Path to package}

\item{excludedPackages}{Packages to exclude}
}
\value{
net_data graph data
}
\description{
getGraphData
}
