% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-getFunctionUse.R, R/summariseFunctionUse.R
\name{getMultiLineFun}
\alias{getMultiLineFun}
\title{getMultiLineFun}
\usage{
getMultiLineFun(line, lines)

getMultiLineFun(line, lines)
}
\arguments{
\item{line}{Current line number}

\item{lines}{All lines}
}
\value{
Returns vector of functions found in do.call function call.

Returns vector of functions found in do.call function call.
}
\description{
getMultiLineFun

getMultiLineFun
}
