% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{getVersionDf}
\alias{getVersionDf}
\title{getVersionDf}
\usage{
getVersionDf(dependencies, permittedPackages)
}
\arguments{
\item{dependencies}{(\link[base]{data.frame})\cr
\enumerate{
  \item (\link[base]{character}) package
  \item (\link[base]{character}) version
}}

\item{permittedPackages}{(\link[base]{data.frame})\cr
\enumerate{
  \item (\link[base]{character}) package
  \item (\link[base]{character}) version
}}
}
\value{
(\link[base]{data.frame})\cr
\enumerate{
  \item (\link[base]{character}) package
  \item (\link[base]{character}) version
}
}
\description{
Function to compare different versions.
}
