% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{getVersionDf}
\alias{getVersionDf}
\title{getVersionDf}
\usage{
getVersionDf(dependencies, permittedPackages)
}
\arguments{
\item{dependencies}{(\link[base]{data.frame})\tabular{ll}{
   column \tab data type \cr
   package \tab \link[base]{character} \cr
   version \tab \link[base]{character} \cr
}}

\item{permittedPackages}{(\link[base]{data.frame})\tabular{ll}{
   column \tab data type \cr
   package \tab \link[base]{character} \cr
   version \tab \link[base]{character} \cr
}}
}
\value{
(\link[base]{data.frame})\tabular{ll}{
   column \tab data type \cr
   package \tab \link[base]{character} \cr
   version \tab \link[base]{character} \cr
}
}
\description{
Function to compare different versions.
}
