% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{getVersionDf}
\alias{getVersionDf}
\title{getVersionDf}
\usage{
getVersionDf(dependencies, permittedPackages)
}
\arguments{
\item{dependencies}{<\link[base]{data.frame}> All dependencies of he package.}

\item{permittedPackages}{<\link[base]{data.frame}> of all permitted packages.}
}
\value{
<\link[base]{data.frame}> with all non permitted packages based on version.
}
\description{
Function to compare different versions.
}
