% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionDiagram.R
\name{graphToDot}
\alias{graphToDot}
\title{graphToDot}
\usage{
graphToDot(graph)
}
\arguments{
\item{graph}{(\link[igraph]{graph})\cr}
}
\value{
`htmlwidgets`\cr
See \link[DiagrammeR]{grViz}.
}
\description{
graphToDot
}
