% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintRepo}
\alias{lintRepo}
\title{lintRepo}
\usage{
lintRepo(repo)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}
}
\value{
<\link[base]{data.frame}> containing lint messages.
}
\description{
lintRepo
}
