% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintRepo}
\alias{lintRepo}
\title{lintRepo}
\usage{
lintRepo(repo)
}
\arguments{
\item{repo}{(`Repository`)
Repository object.}
}
\value{
(`data.frame()`)
  Data.frame containing lint messages.
}
\description{
lintRepo
}
