% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintRepo}
\alias{lintRepo}
\title{lintRepo}
\usage{
lintRepo(repo)
}
\arguments{
\item{repo}{(\link[PaRe]{Repository})\cr}
}
\value{
(\link[base]{data.frame})\cr
\describe{
  \item{filename}{(\link[base]{character}) Name of the file.}
  \item{line_number}{(\link[base]{double}) Line in which the message was found.}
  \item{column_number}{(\link[base]{double}) Column in which the message was found.}
  \item{type}{(\link[base]{character}) Type of message.}
  \item{message}{(\link[base]{character}) Style, warning, or error message.}
  \item{line}{(\link[base]{character}) Line of code in which the message was found.}
  \item{linter}{(\link[base]{character}) Linter used.}
}
}
\description{
Get all the lintr messages of the \link[PaRe]{Repository} object.
}
\examples{
fetchedRepo <- tryCatch({
  # Set dir to clone repository to.
  tempDir <- tempdir()
  pathToRepo <- file.path(tempDir, "glue")

  # Clone repo
  git2r::clone(
    url = "https://github.com/tidyverse/glue.git",
    local_path = pathToRepo)

  # Create instance of Repository object.
  repo <- PaRe::Repository$new(path = pathToRepo)

  # Set fetchedRepo to TRUE if all goes well.
  TRUE
}, error = function(e) {
  # Set fetchedRepo to FALSE if an error is encountered.
  FALSE
}, warning = function(w) {
  # Set fetchedRepo to FALSE if a warning is encountered.
  FALSE
})

if (fetchedRepo) {
  # Run lintRepo on the Repository object.
  messages <- lintRepo(repo = repo)
}
}
