% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lintScore}
\alias{lintScore}
\title{lintScore}
\usage{
lintScore(repo, messages)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}

\item{messages}{<\link[base]{data.frame}> containing lintr messages i.e. from
\link[PaRe]{lintRepo}.}
}
\value{
<\link[dplyr]{tibble}> of percentage scores per type of Lint message.
}
\description{
Function that scores the lintr output as a percentage per message type
(style, warning, error). Lintr messages / lines assessed * 100
}
