% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{makeGraph}
\alias{makeGraph}
\title{makeGraph}
\usage{
makeGraph(funsPerDefFun, pkgName, expFuns, ...)
}
\arguments{
\item{funsPerDefFun}{<\link[base]{data.frame}> Functions per defined function data.frame.}

\item{pkgName}{<\link[base]{character}> Name of package.}

\item{expFuns}{<\link[base]{data.frame}> Exported functinos data.frame.}

\item{...}{Optional other parameters for \link[DiagrammeR]{grViz}.}
}
\value{
<`htmlwidget`> Diagram of the package. See \link[DiagrammeR]{grViz}.
}
\description{
Makes the graph
}
