% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{makeGraph}
\alias{makeGraph}
\title{makeGraph}
\usage{
makeGraph(funsPerDefFun, pkgName, expFuns, ...)
}
\arguments{
\item{funsPerDefFun}{Functions per defined function data.frame.}

\item{pkgName}{Name of package.}

\item{expFuns}{Exported functinos data.frame.}

\item{...}{Optional other parameters for `DiagrammeR::grViz`.}
}
\value{
diagram of the package
}
\description{
Makes the graph
}
