% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{makeReport}
\usage{
makeReport(repo, outputFile, showCode = FALSE)
}
\arguments{
\item{repo}{<\link[PaRe]{Repository}> object.}

\item{outputFile}{<\link[base]{character}> Path to html-file.}

\item{showCode}{<\link[base]{logical}> Default: FALSE; Boolean to show code or not in the report.}
}
\value{
`NULL`
}
\description{
Uses rmarkdown's render function to render a html-report of the given
package.
}
