% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaRe-package.R
\docType{package}
\name{PaRe-package}
\alias{PaRe}
\alias{PaRe-package}
\title{PaRe: A Way to Perform Code Review or QA on Other Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reviews other packages during code review by looking at their dependencies, code style, code complexity, and how internally defined functions interact with one another.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/darwin-eu-dev/PaRe}
  \item Report bugs at \url{https://github.com/darwin-eu-dev/PaRe/issues}
}

}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl} (\href{https://orcid.org/0009-0006-8832-6030}{ORCID})

}
\keyword{internal}
