% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pare.R
\name{pare}
\alias{pare}
\title{A pare Function}
\usage{
pare(labels, predictors, main = "", plot = TRUE)
}
\arguments{
\item{labels}{??? a factor object predictors numeric??? plot default=TRUE to defaultWidth height height of graph defaults to defaultHeight}
}
\description{
This function allows you to plot PARE scores using the ggplot2 package
}
\examples{
pare(4,5,"main",FALSE, width=5,height=100) ???
}
\keyword{accuracy}
\keyword{labels,}
\keyword{predictors,}
