% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiagram.R
\name{pkgDiagram}
\alias{pkgDiagram}
\title{pkgDiagram}
\usage{
pkgDiagram(pkgPath, verbose = FALSE, ...)
}
\arguments{
\item{pkgPath}{Path to package}

\item{verbose}{Verbose messages}

\item{...}{Optional other parameters for `DiagrammeR::grViz`.}
}
\value{
diagram image
}
\description{
Creates a diagram of all defined functions in a package.
}
\examples{
if (interactive()) {
  pkgDiagram(
    pkgPath = "./",
    verbose = TRUE)
}
}
