% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{printMessage}
\alias{printMessage}
\title{printMessage}
\usage{
printMessage(notPermitted, versionCheck)
}
\arguments{
\item{notPermitted}{<[base]{data.frame}> notPermitted}

\item{versionCheck}{<[base]{data.frame}> versionCheck}
}
\value{
<\link[base]{data.frame}> or <\link[base]{NULL}>
}
\description{
Prints messages dependening of the nrow of the number of rows of the
notPermitted and versionCheck data.frames
}
