% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{printMessage}
\alias{printMessage}
\title{printMessage}
\usage{
printMessage(notPermitted, versionCheck)
}
\arguments{
\item{notPermitted}{([base]{data.frame})}

\item{versionCheck}{([base]{data.frame})}
}
\value{
(\link[base]{data.frame})\tabular{ll}{
   column \tab data type \cr
   package \tab \link[base]{character} \cr
   version \tab \link[base]{character} \cr
}
}
\description{
Prints messages dependening of the nrow of the number of rows of the
notPermitted and versionCheck data.frames
}
